requirements.txt
streamlit
huggingface-hub
Pillow
requests
torch
opencv-python
numpy==1.26.4
pytesseract==0.3.10
pyperclip==1.8.2
transformers==4.38.2
pandas==2.2.0
tabulate==0.9.0
easyocr==1.7.1 